STEP 2: Faith
All the scenes you will see in this film are true and are taken only
from life.

If often they are shocking, it is because there are many shocking things
in this world.

Besides, the duty of the chronicler is not to sweeten the truth but to
report it objectively.